% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenario.R
\name{get_scenario}
\alias{get_scenario}
\title{Get Sea Level Rise (SLR) scenario}
\usage{
get_scenario(
  id = 520,
  scenario = c("IntLow", "Int", "IntHigh"),
  method = "curl",
  quiet = TRUE
)
}
\arguments{
\item{id}{numeric, gauge number. Default is 520 (St. Petersburg, FL).}

\item{scenario}{character, SLR scenario. Default is \code{'IntLow'}, \code{'Int'}, and \code{'IntHigh'} (NOAA Intermediate Low, Intermediate, and Intermediate High). See details for available options.}

\item{method}{character, download method. Default is \code{'curl'}, passed to \code{download.file()}.}

\item{quiet}{logical, suppress download messages. Default is \code{TRUE}.}
}
\value{
A data frame with columns for id, scenario, year, SLR in meters, and SLR in feet.
}
\description{
Get Sea Level Rise (SLR) scenario
}
\details{
Information from \url{https://sealevel.nasa.gov/task-force-scenario-tool?psmsl_id=520}, by default. Results are SLR in meters and feet for the intermediate low, intermediate, and intermediate high scenarios based on recommended scenarios from the Climate Science Advisory Panel.  Full options for scenarios are \code{'Low'}, \code{'IntLow'}, \code{'Int'}, \code{'IntHigh'}, and \code{'High'}. Values for SLR are relative change from 2020.
}
\examples{
dat <- get_scenario()
head(dat)
}
