% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sealevel.R
\name{get_sealevel}
\alias{get_sealevel}
\title{Get monthly Mean Sea Level (MSL) data for selected gauge}
\usage{
get_sealevel(gauge = 8726520)
}
\arguments{
\item{gauge}{numeric, gauge number. Default is 8726520 (St. Petersburg, FL).}
}
\value{
A data frame with columns for gauge, year, month, date, MSL in meters, and MSL in feet.
}
\description{
Get monthly Mean Sea Level (MSL) data for selected gauge
}
\details{
Information from \url{https://tidesandcurrents.noaa.gov} using the URL \url{https://tidesandcurrents.noaa.gov/sltrends/data/8726520_meantrend.txt}, by default.  Results are monthly Mean Sea Level (MSL) in meters and feet, with the seasonal cycle removed.
}
\examples{
dat <- get_sealevel(gauge = 8726520)
head(dat)
}
