% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scenario.R
\name{plot_scenario}
\alias{plot_scenario}
\title{Plot sea level scenario data}
\usage{
plot_scenario(
  dat,
  cols = c("deepskyblue", "orange", "red"),
  units = "ft",
  linewidth = 1,
  caption = TRUE,
  title = TRUE,
  subtitle = "Gauge 8726520, St. Petersburg, FL",
  xrng = c(2020, 2100),
  xbrk = 10,
  yrng = c(0, 6),
  ybrk = 1
)
}
\arguments{
\item{dat}{Input data from \code{get_scenario()}.}

\item{cols}{character, color palette for the lines. Default is \code{c('deepskyblue', 'orange', 'red')}.}

\item{units}{character, units for the y-axis. Default is \code{'ft'}. Options are \code{'ft'} and \code{'m'}.}

\item{linewidth}{numeric, line width. Default is \code{1}.}

\item{caption}{logical, add caption with source. Default is \code{TRUE}.}

\item{title}{logical, add title. Default is \code{TRUE}.}

\item{subtitle}{character, subtitle for the plot. Default is \code{'Gauge 8726520, St. Petersburg, FL'}. Use \code{NULL} to omit.}

\item{xrng}{numeric, x-axis range. Default is \code{c(2020, 2100)}.}

\item{xbrk}{numeric, x-axis breaks. Default is \code{10}.}

\item{yrng}{numeric, y-axis range. Default is \code{c(0, 6)}.}

\item{ybrk}{numeric, y-axis breaks. Default is \code{1}.}
}
\value{
a ggplot object
}
\description{
Plot sea level scenario data
}
\examples{
dat <- get_scenario()
plot_scenario(dat)
}
