% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sealevel.R
\name{plot_sealevel}
\alias{plot_sealevel}
\title{Plot sea level data}
\usage{
plot_sealevel(
  dat,
  col = "deepskyblue",
  units = "ft",
  caption = TRUE,
  xrng = NULL,
  xbrk = 10,
  yrng = c(-1, 1),
  ybrk = 0.5
)
}
\arguments{
\item{dat}{Input data from \code{get_sealevel()}.}

\item{col}{character, color for the line. Default is \code{'deepskyblue'}.}

\item{units}{character, units for the y-axis. Default is \code{'ft'}. Options are \code{'ft'} and \code{'m'}.}

\item{caption}{logical, add caption with source. Default is \code{TRUE}.}

\item{xrng}{Date, x-axis range. Default is \code{NULL}, which uses the range of the data.}

\item{xbrk}{numeric, x-axis breaks. Default is \code{10}.}

\item{yrng}{numeric, y-axis range. Default is \code{c(-1, 1)}.}

\item{ybrk}{numeric, y-axis breaks. Default is \code{0.5}.}
}
\value{
a ggplot object
}
\description{
Plot sea level data
}
\examples{
dat <- get_sealevel()
plot_sealevel(dat)
}
