% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scenario.R
\name{plot_scenario}
\alias{plot_scenario}
\title{Plot sea level scenario data}
\usage{
plot_scenario(
  dat,
  cols = c("deepskyblue", "orange", "red"),
  units = "ft",
  linewidth = 1,
  caption = TRUE,
  xrng = c(2020, 2100),
  xbrk = 10,
  yrng = NULL,
  ybrk = 7,
  plotly = FALSE
)
}
\arguments{
\item{dat}{Input data from \code{get_scenario()}.}

\item{cols}{character, color palette for the lines. Default is \code{c('deepskyblue', 'orange', 'red')}.}

\item{units}{character, units for the y-axis. Default is \code{'ft'}. Options are \code{'ft'} and \code{'m'}.}

\item{linewidth}{numeric, line width. Default is \code{1}.}

\item{caption}{logical, add caption with source. Default is \code{TRUE}. Does not apply if \code{plotly = TRUE}.}

\item{xrng}{numeric, x-axis range. Default is \code{c(2020, 2100)}.}

\item{xbrk}{numeric, number of x-axis breaks. Default is \code{10}. Does not apply if \code{plotly = TRUE}.}

\item{yrng}{numeric, y-axis range as two values. Default is \code{NULL}, which uses the range of the data.}

\item{ybrk}{numeric, number of y-axis breaks. Default is \code{7}. Does not apply if \code{plotly = TRUE}.}

\item{plotly}{logical, if \code{TRUE}, returns a plotly object instead of ggplot. Default is \code{FALSE}.}
}
\value{
A ggplot object
}
\description{
Plot sea level scenario data
}
\examples{
dat <- get_scenario()
plot_scenario(dat)
}
